chore: refactor index into Form component
