+++ app/pages/cif-external/project-revision/[projectRevision]/form/[formIndex]/index.tsx
@@ -1,54 +1,23 @@
-/* eslint-disable relay/must-colocate-fragment-spreads*/
-import { Button } from "@button-inc/bcgov-theme";
-import DefaultLayout from "components/Layout/DefaultLayout";
+import Form from "components/Form/Form";
+import ExternalLayout from "components/Layout/ExternalLayout";
+import ExternalTaskList from "components/TaskList/ExternalTaskList";
 import { TaskListMode } from "components/TaskList/types";
-import { useExternalFormPages } from "data/formPages/externalFormStructure";
-import useRedirectTo404IfFalsy from "hooks/useRedirectTo404IfFalsy";
-import useRedirectToLatestRevision from "hooks/useRedirectToLatestRevision";
-import useRedirectToValidFormIndex from "hooks/useRedirectToValidFormIndex";
 import withRelayOptions from "lib/relay/withRelayOptions";
-import { useCreateProjectRevision } from "mutations/ProjectRevision/createProjectRevision";
-import { useRouter } from "next/router";
-import {
-  getExternalProjectRevisionFormPageRoute,
-  getExternalProjectRevisionPageRoute,
-  getProjectRevisionPageRoute,
-} from "routes/pageRoutes";
 import { graphql, usePreloadedQuery } from "react-relay/hooks";
 import { RelayProps, withRelay } from "relay-nextjs";
 import { FormIndexExternalPageQuery } from "__generated__/FormIndexExternalPageQuery.graphql";
-import ExternalTaskList from "components/TaskList/ExternalTaskList";
 
 const pageQuery = graphql`
   query FormIndexExternalPageQuery($projectRevision: ID!) {
     query {
       session {
-        ...DefaultLayout_session
+        ...ExternalLayout_session
       }
       projectRevision(id: $projectRevision) {
-        id
         changeStatus
         projectId
-        projectByProjectId {
-          pendingProjectRevision {
-            id
-          }
-          latestCommittedProjectRevision {
-            id
-          }
-        }
-        projectFormChange {
-          asProject {
-            fundingStreamRfpByFundingStreamRfpId {
-              fundingStreamByFundingStreamId {
-                name
-              }
-            }
-          }
-        }
-
-        ...ApplicationOverviewForm_projectRevision
       }
+      ...Form_query
     }
   }
 `;
@@ -57,136 +26,18 @@ export function ProjectFormPage({
   preloadedQuery,
 }: RelayProps<{}, FormIndexExternalPageQuery>) {
   const { query } = usePreloadedQuery(pageQuery, preloadedQuery);
-  const router = useRouter();
 
   let mode: TaskListMode;
   if (!query.projectRevision?.projectId) mode = "create";
   else if (query.projectRevision.changeStatus === "committed") mode = "view";
   else mode = "update";
 
-  const formIndex = Number(router.query.formIndex);
-
-  const existingRevision =
-    query.projectRevision?.projectByProjectId?.pendingProjectRevision;
-
-  const [createProjectRevision, isCreatingProjectRevision] =
-    useCreateProjectRevision();
-
-  const isRedirecting = useRedirectTo404IfFalsy(query.projectRevision);
-
-  const isRedirectingToLatestRevision = useRedirectToLatestRevision(
-    query.projectRevision?.id,
-    query.projectRevision?.projectByProjectId?.latestCommittedProjectRevision
-      ?.id,
-    mode === "view"
-  );
-
-  const isRedirectingNoFundingStream = useRedirectTo404IfFalsy(
-    query.projectRevision?.projectFormChange?.asProject
-      ?.fundingStreamRfpByFundingStreamRfpId?.fundingStreamByFundingStreamId
-      ?.name
-  );
-
-  const fundingStream = isRedirectingNoFundingStream
-    ? "err"
-    : query.projectRevision.projectFormChange.asProject
-        .fundingStreamRfpByFundingStreamRfpId.fundingStreamByFundingStreamId
-        .name;
-  const formPages = useExternalFormPages(fundingStream);
-
-  const isRedirectingToValidFormIndex = useRedirectToValidFormIndex(
-    formIndex,
-    formPages.length
-  );
-
-  if (
-    isRedirecting ||
-    isRedirectingNoFundingStream ||
-    isRedirectingToLatestRevision ||
-    isRedirectingToValidFormIndex
-  )
-    return null;
-
-  const handleCreateRevision = () => {
-    createProjectRevision({
-      variables: { projectId: query.projectRevision.projectId },
-      onCompleted: (response) => {
-        router.push(
-          getExternalProjectRevisionFormPageRoute(
-            response.createProjectRevision.projectRevision.id,
-            router.query.formIndex as string
-          )
-        );
-      },
-    });
-  };
-  const handleResumeRevision = () => {
-    router.push(
-      getExternalProjectRevisionFormPageRoute(
-        query.projectRevision.projectByProjectId.pendingProjectRevision.id,
-        router.query.formIndex as string
-      )
-    );
-  };
-
-  const createEditButton = () => {
-    return (
-      <div>
-        <Button
-          className="edit-button"
-          onClick={
-            existingRevision ? handleResumeRevision : handleCreateRevision
-          }
-          disabled={isCreatingProjectRevision}
-        >
-          {existingRevision ? "Resume Edition" : "Edit"}
-        </Button>
-        <style jsx>{`
-          div :global(.edit-button) {
-            float: right;
-          }
-        `}</style>
-      </div>
-    );
-  };
-
   const taskList = <ExternalTaskList />;
 
-  const handleSubmit = () => {
-    if (mode === "update" || formIndex === formPages.length - 1) {
-      router.push(
-        getExternalProjectRevisionPageRoute(query.projectRevision.id)
-      );
-    } else {
-      router.push(
-        getExternalProjectRevisionFormPageRoute(
-          query.projectRevision.id,
-          formIndex + 1
-        )
-      );
-    }
-  };
-
-  const EditComponent = formPages[formIndex].editComponent;
-  const ViewComponent = formPages[formIndex].viewComponent;
   return (
-    <DefaultLayout session={query.session} leftSideNav={taskList}>
-      {query.projectRevision.changeStatus === "committed" && ViewComponent ? (
-        <>
-          {createEditButton()}
-          <ViewComponent
-            projectRevision={query.projectRevision}
-            viewOnly={true}
-          />
-        </>
-      ) : (
-        <EditComponent
-          query={query}
-          projectRevision={query.projectRevision}
-          onSubmit={handleSubmit}
-        />
-      )}
-    </DefaultLayout>
+    <ExternalLayout session={query.session} leftSideNav={taskList}>
+      <Form query={query} mode={mode} isInternal={false} />
+    </ExternalLayout>
   );
 }
