+++ app/routes/pageRoutes.ts
@@ -61,21 +61,29 @@ export const getNewProjectRevisionPageRoute = () => ({
   pathname: `/cif/project-revision/new`,
 });
 
-export const getProjectRevisionPageRoute = (projectRevisionId: string) => ({
-  pathname: `/cif/project-revision/[projectRevision]/`,
-  query: {
-    projectRevision: projectRevisionId,
-  },
-});
+export const getProjectRevisionPageRoute = (
+  projectRevisionId: string,
+  isInternal: boolean = true
+) => {
+  const userType = isInternal ? "/cif" : "/cif-external";
+  return {
+    pathname: `${userType}/project-revision/[projectRevision]/`,
+    query: {
+      projectRevision: projectRevisionId,
+    },
+  };
+};
 
 export const getProjectRevisionFormPageRoute = (
   projectRevisionId: string,
   formIndex: string | number,
   anchor: string = undefined,
+  isInternal: boolean = true,
   isRoutedFromNew?: boolean
 ) => {
+  const userType = isInternal ? "/cif" : "/cif-external";
   const urlObject: TaskListLinkUrl = {
-    pathname: `/cif/project-revision/[projectRevision]/form/[formIndex]`,
+    pathname: `${userType}/project-revision/[projectRevision]/form/[formIndex]`,
     query: {
       projectRevision: projectRevisionId,
       formIndex,
